chore: refactor method updatemultiple
